fix: Validation UserService fixed

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDTO } from './dto/create-user.dto';
 import { UpdateUserDTO } from './dto/update-user.dto';
@@ -16,18 +16,36 @@ export class UsersService {
     }
 
     async findById(id: number) {
+        const user = await this.prisma.user.findUnique({where: {id}})
+        
+        if(!user){
+            throw new NotFoundException('User not found')
+        }
+        
         return this.prisma.user.findUnique({where: {id}, 
             select: {id: true, name: true, email: true, password: false, createdAt: true, updatedAt: true}
         })
     }
 
     async update(id: number, dto: UpdateUserDTO) {
+        const user = await this.prisma.user.findUnique({where: {id}})
+        
+        if(!user){
+            throw new NotFoundException('User not found')
+        }
+
         return this.prisma.user.update({where: {id}, data: {name: dto.name}, 
             select: {id: true, name: true, email: true, password: false, createdAt: true, updatedAt: true}
         })
     }
 
     async deleteById(id: number) {
+        const user = await this.prisma.user.findUnique({where: {id}})
+        
+        if(!user){
+            throw new NotFoundException('User not found')
+        }
+
         return this.prisma.user.delete({where: {id}, select: {id: true, name: true, email: true, password: false}})
     }
 }
